+++ src/animations/fadeIn.js
@@ -1,0 +1,8 @@
+import { keyframes } from 'styled-components';
+
+const fadeIn = () => keyframes`
+  0% { opacity: 0; }
+  100% { opacity: 1; }
+`;
+
+export default fadeIn;