+++ src/animations/index.js
@@ -1,3 +1,4 @@
 export fullCircleRotate from './fullCircleRotate';
 export pulseInfinite from './pulseInfinite';
 export slideIn from './slideIn';
+export fadeIn from './fadeIn';