+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Added
+- ToolTip component
+
 ## [3.3.0] - 2018-12-24
 
 ### Changed